requirements.txt
kubernetes
notion-client==2.3.0
python-dotenv==1.0.1
requests==2.32.3